+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Constants/ResetSettingsOptions.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GraphQLAPI\GraphQLAPI\Constants;
+
+class ResetSettingsOptions
+{
+    public final const SAFE = 'safe';
+    public final const UNSAFE = 'unsafe';
+}