+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/PluginEnvironment.php
@@ -5,13 +5,12 @@ declare(strict_types=1);
 namespace GraphQLAPI\GraphQLAPI;
 
 use GraphQLAPI\GraphQLAPI\StaticHelpers\PluginEnvironmentHelpers;
-use PoP\Root\Environment as RootEnvironment;
 
 class PluginEnvironment
 {
     public final const DISABLE_CONTAINER_CACHING = 'DISABLE_CONTAINER_CACHING';
     public final const CACHE_DIR = 'CACHE_DIR';
-    public final const ENABLE_UNSAFE_DEFAULTS = 'ENABLE_UNSAFE_DEFAULTS';
+    public final const SETTINGS_OPTION_ENABLE_UNSAFE_DEFAULT_BEHAVIOR = 'SETTINGS_OPTION_ENABLE_UNSAFE_DEFAULT_BEHAVIOR';
 
     /**
      * If the information is provided by either environment variable
@@ -53,16 +52,24 @@ class PluginEnvironment
         // return dirname(__FILE__, 2) . \DIRECTORY_SEPARATOR . 'cache';
     }
 
-    public static function areUnsafeDefaultsEnabled(): bool
+    public static function getDefinedEnableUnsafeDefaults(): ?bool
     {
-        if (getenv(self::ENABLE_UNSAFE_DEFAULTS) !== false) {
-            return (bool)getenv(self::ENABLE_UNSAFE_DEFAULTS);
+        /**
+         * Priority to decide:
+         *
+         * 1. If env var `SETTINGS_OPTION_ENABLE_UNSAFE_DEFAULT_BEHAVIOR` is defined
+         */
+        if (getenv(self::SETTINGS_OPTION_ENABLE_UNSAFE_DEFAULT_BEHAVIOR) !== false) {
+            return (bool)getenv(self::SETTINGS_OPTION_ENABLE_UNSAFE_DEFAULT_BEHAVIOR);
         }
 
-        if (PluginEnvironmentHelpers::isWPConfigConstantDefined(self::ENABLE_UNSAFE_DEFAULTS)) {
-            return (bool)PluginEnvironmentHelpers::getWPConfigConstantValue(self::ENABLE_UNSAFE_DEFAULTS);
+        /**
+         * 2. If wp-config.php constant `GRAPHQL_API_SETTINGS_OPTION_ENABLE_UNSAFE_DEFAULT_BEHAVIOR` is defined
+         */
+        if (PluginEnvironmentHelpers::isWPConfigConstantDefined(self::SETTINGS_OPTION_ENABLE_UNSAFE_DEFAULT_BEHAVIOR)) {
+            return (bool)PluginEnvironmentHelpers::getWPConfigConstantValue(self::SETTINGS_OPTION_ENABLE_UNSAFE_DEFAULT_BEHAVIOR);
         }
 
-        return RootEnvironment::isApplicationEnvironmentDev();
+        return null;
     }
 }